15 update simluation orchestration and add integration tests

+++ mathhammer/internal/units/units.go
@@ -5,7 +5,7 @@ type Attacker struct {
 	BS        int
 	Strength  int
 	AP        int
-	Damage    int
+	Damage    float32
 	Sustained int
 	//No rerolls = 0, re-roll 1s = 1, re-roll all failed = 6
 	HitReRolls   int
